requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
redis==6.2.0
openai==1.57.0
pydantic==2.11.0
python-multipart==0.0.12
python-dotenv==1.0.0
